use build flag to build smoke_test targets
